Second Refactor
